Add ability to override function name and schema to #[pg_extern]
This commit adds the ability to override the SQL name of a function
defined using pg_extern. Though Postgres allows multiple functions to
have the same name, we currently cannot define such function families
using pgx because the externally-generated symbol created for the rust
function must be globally unique. This commit adds a field to the
#[pg_extern] macro, #[pg_extern(name = "foo")] allowing users to define
a different SQL name for a function than the rust name. For symmetry,
the commit adds a similar override for schema.